requirements.txt
# Core dependencies with version constraints to prevent httpx compatibility issues
openai>=1.99.9,<2.0.0
anthropic>=0.21.0
httpx>=0.27.0,<0.28.0  # Pin to prevent proxies argument conflict

# Other dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.60.0

# Observability dependencies
psutil>=5.9.0
flask>=2.3.0
flask-cors>=4.0.0

# Image generation dependencies
replicate>=0.25.0
Pillow>=10.0.0
requests>=2.31.0